@@ -1,20 +1,24 @@
 import React from "react"
-import ReactDOM from "react-dom"
+import { createRoot } from "react-dom/client"
 import { BrowserRouter } from "react-router-dom"
 import { SocketContext, socket } from "./context/socket"
 
-//Styles
+// Styles
 import "bootstrap/dist/css/bootstrap.min.css"
 import "./style.css"
 
-//Components
+// Components
 import App from "./App"
 
-ReactDOM.render(
-  <SocketContext.Provider value={socket}>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
-  </SocketContext.Provider>,
-  document.getElementById("root")
+const container = document.getElementById("root")
+const root = createRoot(container)
+
+root.render(
+  <React.StrictMode>
+    <SocketContext.Provider value={socket}>
+      <BrowserRouter>
+        <App />
+      </BrowserRouter>
+    </SocketContext.Provider>
+  </React.StrictMode>
 )
